Moved PagePhones and CatalogLayout to components folder.
