+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import PhonesPage from './pages/Phones';
 import PageNotFound from './pages/PageNotFound';
 import { Header } from './components/Header/Header';
 import './styles/_reset.scss';
-import { CatalogLayout } from './pages/CatalogLayout';
+import { CatalogLayout } from './components/Layouts/CatalogLayout';
 import Footer from './components/Footer/Footer';
+import PagePhones from './components/PagePhones/PagePhones';
 
 const App: React.FC = () => {
   return (
@@ -14,7 +14,7 @@ const App: React.FC = () => {
 
       <Routes>
         <Route element={<CatalogLayout />}>
-          <Route path="/phones" element={<PhonesPage />} />
+          <Route path="/phones" element={<PagePhones />} />
         </Route>
         <Route path="*" element={<PageNotFound />} />
       </Routes>
