+++ src/components/Layouts/layouts.scss
@@ -1,4 +1,4 @@
-@import '../utils/mixins.scss';
+@import '../../utils/mixins.scss';
 
 .layout_catalog_container {
   display: flex;
