+++ src/components/PagePhones/PagePhones.scss
@@ -1,6 +1,6 @@
-@import '../utils/fonts.scss';
-@import '../utils/mixins.scss';
-@import '../utils/variables.scss';
+@import '../../utils/fonts.scss';
+@import '../../utils/mixins.scss';
+@import '../../utils/variables.scss';
 
 .phones {
   width: 100%;
